requirements.txt
Flask
requests
readability-lxml
gunicorn